requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.12.0
google-search-results==2.4.2
reportlab==4.0.4
requests==2.32.3